backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dnspython==2.7.0
iniconfig==2.1.0
kombu==5.5.2
lxml==5.3.2
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pymongo==4.12.0
PyPDF2==3.0.1
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
requests==2.31.0
openai
